Store and sqlite data layer
The data layer using Sqlite as the database created. Store added to Server struct and DB connection pool opened when the webserver starts.
