+++ internal/store/sqlite/sqlite.go
@@ -1,0 +1,18 @@
+package sqlite
+
+import (
+	"database/sql"
+	"fmt"
+)
+
+type store struct {
+	DB *sql.DB
+}
+
+func NewStore(DB *sql.DB) *store {
+	return &store{DB: DB}
+}
+
+func (s store) Example() {
+	fmt.Println("example")
+}
